pyproject.toml
[project]
name = "antenna"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "deepl>=1.22.0",
    "feedparser>=6.0.11",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[project.scripts]
antenna = "antenna.__main__:main"
